requirements.txt
streamlit
PyPDF2
pypdf
langchain
langchain-community
langchain_core
langchain-groq
python-dotenv
openai
tiktoken
pinecone-client
streamlit-chat==0.0.2.2
sentence-transformers
